Data preparation is hopefully done

+++ Part_2/prepNoSQL.js
@@ -1,7 +1,9 @@
 var dbs = db.getMongo().getDBNames()
 for(var i in dbs){
     db = db.getMongo().getDB( dbs[i] );
-    print( "dropping db " + db.getName() );
-    db.dropDatabase();
+    if(db.getName() == 'corona'){
+        print("dropping " + db.getName() );
+        db.dropDatabase();
+    }
 }
